tired to fix round issue with expense values below 1 not printing

+++ Team1TravelExpenseApp/src/ca/ualberta/cs/team1travelexpenseapp/Expense.java
@@ -221,7 +221,7 @@ public class Expense {
 		//if(getDescription().length() > 0){
 			str += "\nDescription:" + getDescription();
 		//}
-		if (getAmount().intValue() != 0){
+		if (getAmount().floatValue() != 0){
 			str += "\n" + getAmount() + getCurrency();
 		}
 		if ( receipt != null) {
